renamed package to org.ossgang
