+++ gradle.properties
@@ -1,4 +1,4 @@
-POM.groupId=io.github.ossgang
+POM.groupId=org.ossgang
 POM.artifactId=ossgang-commons
 POM.description=Common Java Utility Classes
