--- src/main/java/io/github/ossgang/commons/observable/Observables.java
@@ -1,28 +1,0 @@
-package io.github.ossgang.commons.observable;
-
-import java.util.Optional;
-
-/**
- * Static support class for dealing with {@link Observable} and {@link ObservableValue}.
- */
-public class Observables {
-    private Observables() {
-        throw new UnsupportedOperationException("static only");
-    }
-
-    /**
-     * Create an {@link ObservableValue} from any {@link Observable}. If the observable passed in is an instance of
-     * {@link ObservableValue}, it is returned unchanged. Otherwise, a derived observable value is created which
-     * subscribes to the upstream observable and caches its latest value.
-     *
-     * @param observable the observable
-     * @param <T> the type
-     * @return an {@link ObservableValue} reflecting the observable provided
-     */
-    public static <T> ObservableValue<T> observableValueOf(Observable<T> observable) {
-        if (observable instanceof ObservableValue) {
-            return (ObservableValue<T>) observable;
-        }
-        return new DerivedObservableValue<>(observable, null, Optional::of);
-    }
-}
