--- src/main/java/io/github/ossgang/commons/property/WrapperProperty.java
@@ -1,55 +1,0 @@
-// @formatter:off
-/*******************************************************************************
- *
- * This file is part of ossgang-commons.
- *
- * Copyright (c) 2008-2019, CERN. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- ******************************************************************************/
-// @formatter:on
-
-package io.github.ossgang.commons.property;
-
-import io.github.ossgang.commons.observable.ObservableValue;
-import io.github.ossgang.commons.observable.Observer;
-import io.github.ossgang.commons.observable.Subscription;
-import io.github.ossgang.commons.observable.SubscriptionOption;
-
-import java.util.function.Consumer;
-
-public class WrapperProperty<T> implements Property<T>, ObservableValue<T> {
-    private final ObservableValue<T> observableValue;
-    private final Consumer<? super T> setConsumer;
-
-    WrapperProperty(ObservableValue<T> updateProvider, Consumer<? super T> setConsumer) {
-        this.observableValue = updateProvider;
-        this.setConsumer = setConsumer;
-    }
-
-    @Override
-    public void set(T value) {
-        setConsumer.accept(value);
-    }
-
-    @Override
-    public T get() {
-        return observableValue.get();
-    }
-
-    @Override
-    public Subscription subscribe(Observer<? super T> listener, SubscriptionOption... options) {
-        return observableValue.subscribe(listener, options);
-    }
-}
